Add files via upload
Muon trajectory projections

The geogebra file is for displaying the interactions of the geometry itself given R0, R1, R_solenoid, and R_Station_0. The muon follows R0 to the solenoid, then R1 to Station_0.

The mathematica file calculates these Radii using the relativistic gyroradius formula which can be found on wikipedia: https://en.wikipedia.org/wiki/Gyroradius. Units in the code match units in the wikipedia page. Note that R0 and R1 are changed to R1 and R2 in mathematica code.